Move coalesce out of spark_only into operations
Reviewer asked whether dsgrid/ibis/spark_only.py should keep existing
as a one-function module. Folding coalesce into operations.py and
deleting spark_only.py keeps the partition-style helpers in one place
and removes a discovery hop for callers.

No behavior change: coalesce remains a DuckDB no-op and still
re-registers a coalesced PySpark DataFrame on Spark. Five import sites
updated.
